refactor: Phase 1 - テスト基盤の統合とモック機能の統一
