+++ src/frontend/tests/setup.ts
@@ -7,30 +7,40 @@ afterEach(() => {
   cleanup();
 });
 
+// matchMediaのモック
+Object.defineProperty(window, 'matchMedia', {
+  writable: true,
+  value: vi.fn().mockImplementation((query) => ({
+    matches: false,
+    media: query,
+    onchange: null,
+    addListener: vi.fn(), // Deprecated
+    removeListener: vi.fn(), // Deprecated
+    addEventListener: vi.fn(),
+    removeEventListener: vi.fn(),
+    dispatchEvent: vi.fn(),
+  })),
+});
+
+// ResizeObserverのモック
+global.ResizeObserver = vi.fn().mockImplementation(() => ({
+  observe: vi.fn(),
+  unobserve: vi.fn(),
+  disconnect: vi.fn(),
+}));
+
+// DOM scrollTo メソッドのモック
+Object.defineProperty(HTMLDivElement.prototype, 'scrollTo', {
+  value: vi.fn(),
+  writable: true,
+});
+
+Object.defineProperty(HTMLDivElement.prototype, 'scrollHeight', {
+  value: 1000,
+  writable: true,
+});
+
 // グローバルモック設定
 vi.mock('nanoid', () => ({
   nanoid: () => 'test-id-12345',
 }));
-
-// WebSocketのグローバルモック
-const MockWebSocket = vi.fn().mockImplementation((url: string) => ({
-  url,
-  readyState: WebSocket.CONNECTING,
-  onopen: null,
-  onclose: null,
-  onmessage: null,
-  onerror: null,
-  send: vi.fn(),
-  close: vi.fn(),
-}));
-
-// WebSocket定数をMockWebSocketに追加
-MockWebSocket.CONNECTING = 0;
-MockWebSocket.OPEN = 1;
-MockWebSocket.CLOSING = 2;
-MockWebSocket.CLOSED = 3;
-
-vi.stubGlobal('WebSocket', MockWebSocket);
-
-// fetchのグローバルモック
-global.fetch = vi.fn();
